fix(TextUtils): Add utf8 decode
Modify TextUtils.stdDecode() to accept an optional boolean parameter isGBK indicating whether the input is encoded in GBK or UTF-8. If isGBK is true, the method uses the gbk.decode() function to convert the input from byte array to a string, followed by removing escape sequences and extra line breaks with removeEscapeSequences() and removeExtraBreaks(). Otherwise, the method uses utf8.decode() to decode the input.

https://github.com/iiijam/alisthelper/issues/1